Add author check

+++ src/main/java/com/mcmoddev/mmdbot/modules/logging/misc/ScamDetector.java
@@ -28,6 +28,7 @@ import com.mcmoddev.mmdbot.MMDBot;
 import com.mcmoddev.mmdbot.utilities.Utils;
 import com.mcmoddev.mmdbot.utilities.console.MMDMarkers;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
@@ -44,10 +45,10 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
+import java.util.HashSet;
 import java.util.Locale;
+import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.StreamSupport;
 
@@ -60,7 +61,7 @@ public class ScamDetector extends ListenerAdapter {
     public static final Gson GSON = new GsonBuilder().disableHtmlEscaping().create();
     public static final String SCAM_LINKS_DATA_URL = "https://phish.sinking.yachts/v2/all";
 
-    public static final List<String> SCAM_LINKS = Collections.synchronizedList(new ArrayList<>());
+    public static final Set<String> SCAM_LINKS = Collections.synchronizedSet(new HashSet<>());
 
     static {
         new Thread(ScamDetector::setupScamLinks, "Scam link collector").start();
@@ -69,8 +70,12 @@ public class ScamDetector extends ListenerAdapter {
     @Override
     public void onGuildMessageReceived(@NotNull final GuildMessageReceivedEvent event) {
         final var msg = event.getMessage();
+        final var member = msg.getMember();
+        if (member == null || msg.getAuthor().isBot() || msg.getAuthor().isSystem() ||
+            member.hasPermission(Permission.MANAGE_CHANNEL)) {
+            return;
+        }
         if (containsScam(msg)) {
-            final var member = msg.getMember();
             final var guild = msg.getGuild();
             final var embed = getLoggingEmbed(msg, "");
             msg.delete().queue($ -> {
@@ -83,8 +88,12 @@ public class ScamDetector extends ListenerAdapter {
     @Override
     public void onGuildMessageUpdate(@NotNull final GuildMessageUpdateEvent event) {
         final var msg = event.getMessage();
+        final var member = msg.getMember();
+        if (member == null || msg.getAuthor().isBot() || msg.getAuthor().isSystem() ||
+            member.hasPermission(Permission.MANAGE_CHANNEL)) {
+            return;
+        }
         if (containsScam(msg)) {
-            final var member = msg.getMember();
             final var guild = msg.getGuild();
             final var embed = getLoggingEmbed(msg, ", by editing an old message");
             msg.delete().queue($ -> {
@@ -123,9 +132,11 @@ public class ScamDetector extends ListenerAdapter {
 
     public static boolean containsScam(final Message message) {
         final String msgContent = message.getContentRaw().toLowerCase(Locale.ROOT);
-        for (final var link : SCAM_LINKS) {
-            if (msgContent.contains(link)) {
-                return true;
+        synchronized (SCAM_LINKS) {
+            for (final var link : SCAM_LINKS) {
+                if (msgContent.contains(link)) {
+                    return true;
+                }
             }
         }
         return false;
